Fix potential compiler warnings.
Backported from branch-1.3-porting.


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@11954 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fluid/fluid.cxx
@@ -130,7 +130,7 @@ void goto_source_dir() {
   if (p <= filename) return; // it is in the current directory
   char buffer[FL_PATH_MAX];
   strlcpy(buffer, filename, sizeof(buffer));
-  int n = p-filename; if (n>1) n--; buffer[n] = 0;
+  int n = (int)(p-filename); if (n>1) n--; buffer[n] = 0;
   if (!pwd) {
     pwd = getcwd(0,FL_PATH_MAX);
     if (!pwd) {fprintf(stderr,"getwd : %s\n",strerror(errno)); return;}
@@ -1773,7 +1773,7 @@ int main(int argc,char **argv) {
       " -cs : write .cxx and .h and strings and exit\n"
       " -o <name> : .cxx output filename, or extension if <name> starts with '.'\n"
       " -h <name> : .h output filename, or extension if <name> starts with '.'\n";
-    int len = strlen(msg) + strlen(argv[0]) + strlen(Fl::help);
+    int len = (int)(strlen(msg) + strlen(argv[0]) + strlen(Fl::help));
     Fl_Plugin_Manager pm("commandline");
     int i, n = pm.plugins();
     for (i=0; i<n; i++) {
